third_party/WebKit/LayoutTests/fast/dom/MutationObserver/observer-wrapper-dropoff-expected.txt

MutationObserver wrappers should survive GC for passing into the callback even if JS has lost references.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS observer.testProperty is true
PASS successfullyParsed is true

TEST COMPLETE
